# query/docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API Server Classroom service",
        "title": "Classroom API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/c/classroom",
    "paths": {
...
        "/forms/get_form/{idForm}": {
            "get": {
                "description": "Get a single user form, ROLS=[teacher]",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "forms",
                    "classroom",
                    "roles.teacher"
                ],
                "summary": "Get a single form",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/res.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/smaps.FormMap"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized role",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "404": {
                        "description": "No existe este formulario",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable - NATS || DB Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
        "/forms/get_forms": {
            "get": {
                "description": "Get forms generated for user",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "forms",
                    "classroom",
                    "roles.teacher"
                ],
                "summary": "Get forms of user, ROLES=[teacher]",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/res.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/smaps.FormsMap"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized role",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable - NATS || DB Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
...
        "/grades/download_grades": {
            "get": {
                "description": "Get student grades (in a semester) -\u003e export to PDF",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/pdf"
                ],
                "tags": [
                    "grades",
                    "classroom",
                    "roles.student",
                    "roles.student_directive"
                ],
                "summary": "Export student grades",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MongoID",
                        "name": "semester",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthorized role",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable - NATS || DB Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "510": {
                        "description": "Buffer io.Writter",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
        "/grades/export_grades/{idModule}": {
            "get": {
                "description": "Get student grades (in a module) -\u003e export to Excel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
                ],
                "tags": [
                    "grades",
                    "classroom",
                    "roles.teacher",
                    "roles.directive",
                    "roles.director"
                ],
                "summary": "Export grades",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Mongo ID Form",
                        "name": "idModule",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthorized role",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable - NATS || DB Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "510": {
                        "description": "Buffer io.Writter",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
        "/grades/get_grade_programs/{idModule}": {
            "get": {
                "description": "Get program grade by MongoId (Sort by program number)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "grades",
                    "classroom",
                    "roles.all"
                ],
                "summary": "Get program grade",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Mongo ID Form",
                        "name": "idModule",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/res.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/smaps.ProgramGradeMap"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable - NATS || DB Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
        "/grades/get_student_grades/{idModule}": {
            "get": {
                "description": "Get student grades (in a module)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "grades",
                    "classroom",
                    "roles.student",
                    "roles.student_directive"
                ],
                "summary": "Get student grades",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Mongo ID Form",
                        "name": "idModule",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthorized role",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable - NATS || DB Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
        "/grades/get_students_grades/{idModule}": {
            "get": {
                "description": "Get all students and their grades (in a module)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "grades",
                    "classroom",
                    "roles.teacher",
                    "roles.director",
                    "roles.directive"
                ],
                "summary": "Get students grades",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Mongo ID Form",
                        "name": "idModule",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthorized role",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable - NATS || DB Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Acumulative": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string",
                    "example": "637850aafa8118c9f0f1202a"
                },
                "number": {
                    "type": "integer",
                    "example": 1
                },
                "percentage": {
                    "type": "number",
                    "example": 20
                }
            }
        },
...
        "models.Form": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string",
                    "example": "6376c8283cc695e19d785b08"
                },
                "author": {
                    "type": "string",
                    "example": "6376c8283cc695e19d785b08"
                },
                "has_points": {
                    "type": "boolean",
                    "example": true
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.FormItem"
                    }
                },
                "status": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string",
                    "example": "This is a form!"
                },
                "update_date": {
                    "type": "string",
                    "example": "2022-09-21T20:10:23.309+00:00"
                },
                "upload_date": {
                    "type": "string",
                    "example": "2022-09-21T20:10:23.309+00:00"
                }
            }
        },
...
        "models.FormItem": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string",
                    "example": "6376c8283cc695e19d785b08"
                },
                "points_type": {
                    "type": "string",
                    "x-omitempty": true
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "6376c8283cc695e19d785b08"
                    ]
                },
                "title": {
                    "type": "string",
                    "example": "This is a item form!"
                }
            }
        },
        "models.FormItemWLookup": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string",
                    "example": "6376c8283cc695e19d785b08"
                },
                "points_type": {
                    "type": "string",
                    "x-omitempty": true
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ItemQuestion"
                    }
                },
                "title": {
                    "type": "string",
                    "example": "This is a item title!"
                }
            }
        },
        "models.FormWLookup": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string",
                    "example": "6376c8283cc695e19d785b08"
                },
                "has_points": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.FormItemWLookup"
                    }
                },
                "title": {
                    "type": "string",
                    "example": "This is a form!"
                },
                "update_date": {
                    "type": "string",
                    "example": "2022-09-21T20:10:23.309+00:00"
                },
                "upload_date": {
                    "type": "string",
                    "example": "2022-09-21T20:10:23.309+00:00"
                }
            }
        },
        "models.GradesProgram": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string",
                    "example": "637850aafa8118c9f0f1202a"
                },
                "acumulative": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Acumulative"
                    },
                    "x-omitempty": true
                },
                "is_acumulative": {
                    "type": "boolean"
                },
                "module": {
                    "type": "string",
                    "example": "637850aafa8118c9f0f1202a"
                },
                "number": {
                    "type": "integer",
                    "example": 3
                },
                "percentage": {
                    "type": "number",
                    "example": 25
                }
            }
        },
        "models.ItemQuestion": {
            "description": "Points is null if (father)form.has_points = false Answers is null if question.type=written Correct is null if question.type=written || alternatives",
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string",
                    "example": "6376c8283cc695e19d785b08"
                },
                "answers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "x-omitempty": true,
                    "example": [
                        "a",
                        " b"
                    ]
                },
                "correct": {
                    "type": "integer",
                    "x-omitempty": true,
                    "example": 2
                },
                "points": {
                    "type": "integer",
                    "x-omitempty": true,
                    "example": 25
                },
                "question": {
                    "type": "string",
                    "example": "Whats your name?"
                },
                "type": {
                    "type": "string",
                    "example": "alternatives_correct"
                }
            }
        },
...
        "res.Response": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "object",
                    "additionalProperties": true
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "smaps.FormMap": {
            "type": "object",
            "properties": {
                "form": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.FormWLookup"
                    }
                }
            }
        },
...
        "smaps.FormsMap": {
            "type": "object",
            "properties": {
                "forms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Form"
                    }
                }
            }
        },
...
        "smaps.ProgramGradeMap": {
            "type": "object",
            "properties": {
                "programs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GradesProgram"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "BearerJWTToken in Authorization Header",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}